add `notdeft-new-file-data-function' option
That is, add a new configuration variable for providing more control
over the creation of new notes. Notably, any function stored in that
variable is able to decide on the file name also for notes created
without a title, and it can also modify the note content before it gets
saved into a file.

Have the default implementation of that function be mostly backward
compatible, and in particular have it still call
`notdeft-title-to-notename' internally so that any existing
title-to-notename mappings are honored.

As a minor breaking change have `notdeft-create-file' no longer support
(format STRING)s for generating file names.