add `notdeft-select-note-file-function' option
This option is for choosing how to select a note file.

It is initially honored by `notdeft-org-link-existing-note'.

The default method of selecting from among all note names with Ido can
get unwieldy if there are great many notes. In that case a ranked search
probably works better even though there then is the extra step of
prompting for a search query. Predefine a
`notdeft-search-select-note-file' function for selecting notes like
that, which can be configured for use with

  (setq notdeft-select-note-file-function
       'notdeft-search-select-note-file)