add a `notdeft-select-note-file-by-search' flag
Add it as a configuration option, as a simpler alternative to
`notdeft-select-note-file-function'.

Change the signature of `notdeft-select-note-file-function' to accept
selection options as an argument. Use one of those options in
`notdeft-org-link-existing-note' to specify the text of any current
region as a potential search string. Modify `notdeft-xapian-read-query'
to permit such a string to be optionally specified.