add `notdeft-cache-compaction-factor'
That is, support optional automatic recollection of file information
cache entries as the cache starts to grow large. This only applies to
the Xapian case, where information is only needed about files appearing
in result sets.

Doing the recollection starts to make sense as with the Xapian backend
NotDeft can otherwise cope with very large sets of files, and it may no
longer be appropriate to indefinitely store the contents of any and all
of them in memory. In the non-Xapian case no such compaction is
possible, as all the note files are in the "result set" at all times.