optionally recognize Org PROPERTIES drawers
That is, when parsing note headers recognize Org mode :PROPERTIES:
drawers and treat them the same as comments. This means that parsing of
a header does not end when such a drawer is encountered, so that for
example a TITLE appearing afterwards will be recognized as such. Also,
the drawer will then not end up in a note summary in a *NotDeft* buffer.

As not everyone uses Org let alone buffer-level :PROPERTIES:, make this
behavior optional, possible to enable by setting the variable
`notdeft-allow-org-property-drawers' to a non-nil value.

Closes: #19